+++ lib/server_settings/host.rb
