+++ lib/server_settings/host_collection.rb
