+++ lib/server_settings/version.rb
